fix: moving more of the mechanics out of the view layer
