+++ src/notebooks/components/header/Submit.tsx
@@ -7,7 +7,6 @@ import {ResultsContext} from 'src/notebooks/context/results'
 import {TimeContext} from 'src/notebooks/context/time'
 import {IconFont} from '@influxdata/clockface'
 import {notify} from 'src/shared/actions/notifications'
-import {PIPE_DEFINITIONS} from 'src/notebooks'
 
 // Utils
 import {event} from 'src/cloud/utils/reporting'
@@ -15,12 +14,10 @@ import {event} from 'src/cloud/utils/reporting'
 // Types
 import {RemoteDataState} from 'src/types'
 
-const PREVIOUS_REGEXP = /__PREVIOUS_RESULT__/g
-
 const fakeNotify = notify
 
 export const Submit: FC = () => {
-  const {query} = useContext(QueryContext)
+  const {query, generateMap} = useContext(QueryContext)
   const {id, notebook} = useContext(NotebookContext)
   const {add, update} = useContext(ResultsContext)
   const {timeContext} = useContext(TimeContext)
@@ -44,73 +41,31 @@ export const Submit: FC = () => {
     event('Notebook Submit Button Clicked')
     setLoading(RemoteDataState.Loading)
 
+    notebook.data.allIDs.forEach(pipeID => {
+      notebook.meta.update(pipeID, {loading: RemoteDataState.Loading})
+    })
+
+    const map = generateMap()
+
     Promise.all(
-      notebook.data.allIDs
-        .reduce((stages, pipeID) => {
-          notebook.meta.update(pipeID, {loading: RemoteDataState.Loading})
-          const pipe = notebook.data.get(pipeID)
-
-          let stage = {
-            text: '',
-            instances: [pipeID],
-            requirements: {},
-          }
-
-          const create = (text, loadPrevious) => {
-            if (loadPrevious && stages.length) {
-              stage.requirements = {
-                ...stages[stages.length - 1].requirements,
-                [`prev_${stages.length}`]: stages[stages.length - 1].text,
-              }
-              stage.text = text.replace(
-                PREVIOUS_REGEXP,
-                `prev_${stages.length}`
-              )
-            } else {
-              stage.text = text
-            }
-
-            stages.push(stage)
-          }
-
-          const append = () => {
-            if (stages.length) {
-              stages[stages.length - 1].instances.push(pipeID)
-            }
-          }
-
-          if (PIPE_DEFINITIONS[pipe.type].generateFlux) {
-            PIPE_DEFINITIONS[pipe.type].generateFlux(pipe, create, append)
-          } else {
-            append()
-          }
-
-          return stages
-        }, [])
-        .map(queryStruct => {
-          const queryText =
-            Object.entries(queryStruct.requirements)
-              .map(([key, value]) => `${key} = (\n${value}\n)\n\n`)
-              .join('') + queryStruct.text
-
-          return query(queryText)
-            .then(response => {
-              queryStruct.instances.forEach(pipeID => {
-                forceUpdate(pipeID, response)
-                notebook.meta.update(pipeID, {loading: RemoteDataState.Done})
-              })
+      map.map(stage => {
+        return query(stage.text)
+          .then(response => {
+            stage.instances.forEach(pipeID => {
+              forceUpdate(pipeID, response)
+              notebook.meta.update(pipeID, {loading: RemoteDataState.Done})
             })
-            .catch(e => {
-              queryStruct.instances.forEach(pipeID => {
-                forceUpdate(pipeID, {
-                  error: e.message,
-                })
-                notebook.meta.update(pipeID, {loading: RemoteDataState.Error})
+          })
+          .catch(e => {
+            stage.instances.forEach(pipeID => {
+              forceUpdate(pipeID, {
+                error: e.message,
               })
+              notebook.meta.update(pipeID, {loading: RemoteDataState.Error})
             })
-        })
+          })
+      })
     )
-
       .then(() => {
         event('Notebook Submit Resolved')
