+++ src/notebooks/context/query.tsx
@@ -12,22 +12,32 @@ import {TimeContext} from 'src/notebooks/context/time'
 import {fromFlux as parse} from '@influxdata/giraffe'
 import {event} from 'src/cloud/utils/reporting'
 import {FluxResult} from 'src/types/notebooks'
+import {PIPE_DEFINITIONS} from 'src/notebooks'
+
+interface Stage {
+  text: string
+  instances: string[]
+}
 
 export interface QueryContextType {
   query: (text: string) => Promise<FluxResult>
+  generateMap: () => Stage[]
 }
 
 export const DEFAULT_CONTEXT: QueryContextType = {
   query: () => Promise.resolve({} as FluxResult),
+  generateMap: () => [],
 }
 
 export const QueryContext = React.createContext<QueryContextType>(
   DEFAULT_CONTEXT
 )
 
+const PREVIOUS_REGEXP = /__PREVIOUS_RESULT__/g
+
 type Props = StateProps
 export const QueryProvider: FC<Props> = ({children, variables, org}) => {
-  const {id} = useContext(NotebookContext)
+  const {id, notebook} = useContext(NotebookContext)
   const {timeContext} = useContext(TimeContext)
   const time = timeContext[id]
 
@@ -40,6 +50,58 @@ export const QueryProvider: FC<Props> = ({children, variables, org}) => {
 
     variables.map(v => asAssignment(v))
   }, [variables, time])
+
+  const generateMap = (): Stage[] => {
+    return notebook.data.allIDs
+      .reduce((stages, pipeID) => {
+        const pipe = notebook.data.get(pipeID)
+
+        let stage = {
+          text: '',
+          instances: [pipeID],
+          requirements: {},
+        }
+
+        const create = (text, loadPrevious) => {
+          if (loadPrevious && stages.length) {
+            stage.requirements = {
+              ...stages[stages.length - 1].requirements,
+              [`prev_${stages.length}`]: stages[stages.length - 1].text,
+            }
+            stage.text = text.replace(PREVIOUS_REGEXP, `prev_${stages.length}`)
+          } else {
+            stage.text = text
+          }
+
+          stages.push(stage)
+        }
+
+        const append = () => {
+          if (stages.length) {
+            stages[stages.length - 1].instances.push(pipeID)
+          }
+        }
+
+        if (PIPE_DEFINITIONS[pipe.type].generateFlux) {
+          PIPE_DEFINITIONS[pipe.type].generateFlux(pipe, create, append)
+        } else {
+          append()
+        }
+
+        return stages
+      }, [])
+      .map(queryStruct => {
+        const queryText =
+          Object.entries(queryStruct.requirements)
+            .map(([key, value]) => `${key} = (\n${value}\n)\n\n`)
+            .join('') + queryStruct.text
+
+        return {
+          text: queryText,
+          instances: queryStruct.instances,
+        }
+      })
+  }
 
   const query = (text: string) => {
     const windowVars = getWindowVars(text, vars)
@@ -69,7 +131,9 @@ export const QueryProvider: FC<Props> = ({children, variables, org}) => {
   }
 
   return (
-    <QueryContext.Provider value={{query}}>{children}</QueryContext.Provider>
+    <QueryContext.Provider value={{query, generateMap}}>
+      {children}
+    </QueryContext.Provider>
   )
 }
